requirements.txt
Django>=5.0,<6.0
pandas>=2.0.0
openpyxl>=3.1.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
python-decouple>=3.8
gunicorn>=21.0.0
